+++ biothings/hub/databuild/auto_archive.py
@@ -4,7 +4,7 @@ from functools import partial
 from dateutil import parser as dtparser
 
 from biothings import config as btconfig
-from biothings.utils.manager import BaseManager
+from biothings.hub.manager import BaseManager
 
 logger = btconfig.logger
