+++ biothings/hub/dataload/dumper.py
@@ -33,11 +33,12 @@ import requests
 
 from biothings import config as btconfig
 from biothings.hub import DUMPER_CATEGORY, UPLOADER_CATEGORY, renderer as job_renderer
+from biothings.hub.manager import ResourceError
 from biothings.hub.dataload.uploader import set_pending_to_upload
+from biothings.hub.dataload.manager import BaseSourceManager
 from biothings.utils.common import rmdashfr, timesofar, untarall
 from biothings.utils.hub_db import get_src_dump
 from biothings.utils.loggers import get_logger
-from biothings.utils.manager import BaseSourceManager, ResourceError
 from biothings.utils.parsers import docker_source_info_parser
 
 logging = btconfig.logger
