+++ biothings/hub/dataplugin/manager.py
@@ -20,8 +20,8 @@ from biothings.utils.common import (
 )
 from biothings import config as btconfig
 from biothings.utils.hub_db import get_data_plugin, get_src_dump, get_src_master
-from biothings.utils.manager import BaseSourceManager
 from biothings.hub.dataload import dumper
+from biothings.hub.dataload.manager import BaseSourceManager
 from biothings.hub.dataplugin.assistant import GithubAssistant, LocalAssistant, AssistantException
