+++ biothings/utils/manager.py
@@ -3,8 +3,6 @@ import concurrent.futures
 import copy
 import datetime
 import glob
-import importlib
-import os
 import re
 import threading
 import time
@@ -29,7 +27,6 @@ except ImportError:
 import biothings.hub  # noqa
 from biothings import config
 from biothings.utils.common import get_random_string, sizeof_fmt, timesofar
-from biothings.utils.hub_db import get_src_conn
 
 logger = config.logger
 
@@ -63,7 +60,7 @@ def track(func):
         # just informative, so stringify is just ok there)
         # make sure we can pickle the whole thing (and it's
         innerargs = [str(arg) for arg in innerargs]
-        if type(innerfunc) == partial:
+        if isinstance(innerfunc, partial):
             fname = innerfunc.func.__name__
         elif isinstance(innerfunc, types.MethodType):
             fname = innerfunc.__self__.__class__.__name__
@@ -148,337 +145,6 @@ def norm(value, maxlen):
     if len(value) > maxlen:
         value = "...%s" % value[-maxlen + 3 :]
     return value
-
-
-class UnknownResource(Exception):
-    pass
-
-
-class ResourceError(Exception):
-    pass
-
-
-class ManagerError(Exception):
-    pass
-
-
-class ResourceNotFound(Exception):
-    pass
-
-
-class BaseManager:
-    def __init__(self, job_manager, poll_schedule=None):
-        self.register = {}
-        self.poll_schedule = poll_schedule
-        self.job_manager = job_manager
-        self.clean_stale_status()
-
-    def clean_stale_status(self):
-        """
-        During startup, search for action in progress which would have
-        been interrupted and change the state to "canceled".
-        Ex: some donwloading processes could have been interrupted, at
-        startup, "downloading" status should be changed to "canceled" so
-        to reflect actual state on these datasources.
-        This must be overriden in subclass.
-        """
-
-    def __repr__(self):
-        registered = sorted(list(self.register.keys()))
-        return "<%s [%d registered]: %s>" % (
-            self.__class__.__name__,
-            len(self.register),
-            registered,
-        )
-
-    def __getitem__(self, src_name):
-        try:
-            # as a main-source
-            return self.register[src_name]
-        except KeyError:
-            try:
-                # as a sub-source
-                main, sub = src_name.split(".")
-                srcs = self.register[main]
-                # there can be many uploader for one resource (when each is dealing
-                # with one specific file but upload to the same collection for instance)
-                # so we want to make sure user is aware of this and not just return one
-                # uploader when many are needed
-                # on the other hand, if only one avail, just return it
-                res = [src for src in srcs if src.name == sub]
-                if not res:
-                    raise KeyError(src_name)
-                return res
-            except (ValueError, AttributeError, KeyError):
-                # nope, can't find it...
-                raise KeyError(src_name)
-
-    def poll(self, state, func, col):
-        """
-        Search for source in collection 'col' with a pending flag list
-        containing 'state' and and call 'func' for each document found
-        (with doc as only param)
-        """
-        if not self.poll_schedule:
-            raise ManagerError("poll_schedule is not defined")
-
-        async def check_pending(state):
-            sources = [src for src in col.find({"pending": state}) if isinstance(src["_id"], str)]
-            if sources:
-                logger.info(
-                    "Found %d resources with pending flag %s (%s)",
-                    len(sources),
-                    state,
-                    repr([src["_id"] for src in sources]),
-                )
-            for src in sources:
-                logger.info("Run %s for pending flag %s on source '%s'", func, state, src["_id"])
-                try:
-                    # first reset flag to make sure we won't call func multiple time
-                    col.update({"_id": src["_id"]}, {"$pull": {"pending": state}})
-                    func(src)
-                except ResourceNotFound:
-                    logger.error(
-                        "Resource '%s' has a pending flag set to %s but is not registered in manager",
-                        src["_id"],
-                        state,
-                    )
-
-        return aiocron.crontab(
-            self.poll_schedule,
-            func=partial(check_pending, state),
-            start=True,
-            loop=self.job_manager.loop,
-        )
-
-
-class BaseStatusRegisterer:
-    def load_doc(self, key_name, stage):
-        """
-        Find document using key_name and stage, stage being a
-        key within the document matching a specific process name:
-        Ex: {"_id":"123","snapshot":"abc"}
-            load_doc("abc","snapshot")
-        will return the document. Note key_name is first used to
-        find the doc by its _id.
-        Ex: with another doc {"_id" : "abc", "snapshot" : "somethingelse"}
-            load_doc{"abc","snapshot")
-        will return doc with _id="abc", not "123"
-        """
-        doc = self.collection.find_one({"_id": key_name})
-        if not doc:
-            doc = []
-            bdocs = self.collection.find()
-            for adoc in bdocs:
-                if key_name in adoc.get(stage, {}):
-                    doc.append(adoc)
-            if len(doc) == 1:
-                # we'll just return the single doc
-                # otherwise it's up to the caller to do something with that
-                doc = doc.pop()
-        assert doc, "No document could be found"
-        return doc
-
-    @property
-    def collection(self):
-        """
-        Return collection object used to fetch doc in which we store status
-        """
-        raise NotImplementedError("implement me in sub-class")
-
-    def register_status(self, doc, stage, status, transient=False, init=False, **extra):
-        assert self.collection, "No collection set"
-        # stage: "snapshot", "publish", etc... depending on the what's being done
-        job_info = {
-            "status": status,
-            "step_started_at": datetime.datetime.now().astimezone(),
-            "logfile": self.logfile,
-        }
-        stage_info = {}
-        stage_key = None
-        # register status can be about different stages:
-        stage_info.setdefault(stage, {}).update(extra[stage])
-        stage_key = list(extra[stage].keys())
-        assert len(stage_key) == 1, stage_key
-        stage_key = stage_key.pop()
-        if transient:
-            # record some "in-progress" information
-            job_info["pid"] = os.getpid()
-        else:
-            # only register time when it's a final state
-            job_info["time"] = timesofar(self.ti)
-            t1 = round(time.time() - self.ti, 0)
-            job_info["time_in_s"] = t1
-            stage_info.setdefault(stage, {}).setdefault(stage_key, {}).update(
-                {"created_at": datetime.datetime.now().astimezone()}
-            )
-        if "job" in extra:
-            job_info.update(extra["job"])
-        # since the base is the merged collection, we register info there
-        if init:
-            # init timer for this step
-            self.ti = time.time()
-            self.collection.update({"_id": doc["_id"]}, {"$push": {"jobs": job_info}})
-            # now refresh/sync
-            doc = self.collection.find_one({"_id": doc["_id"]})
-        else:
-            # merge extra at root level
-            doc["jobs"] and doc["jobs"].append(job_info)
-
-            def merge_index_info(target, d):
-                if not isinstance(target, dict):
-                    # previous value wasn't a dict, just replace
-                    target = d
-                elif "__REPLACE__" in d.keys():
-                    d.pop("__REPLACE__")
-                    target = d
-                else:
-                    if status == "success":
-                        # remove 'err' key to avoid merging success results with errors
-                        target.pop("err", None)
-                    for k, v in d.items():
-                        if isinstance(v, dict):
-                            if k in target:
-                                target[k] = merge_index_info(target[k], v)
-                            else:
-                                v.pop("__REPLACE__", None)
-                                # merge v with "nothing" just to make sure to remove any "__REPLACE__"
-                                v = merge_index_info({}, v)
-                                target[k] = v
-                        else:
-                            target[k] = v
-                return target
-
-            doc = merge_index_info(doc, stage_info)
-            self.collection.replace_one({"_id": doc["_id"]}, doc)
-
-
-class BaseSourceManager(BaseManager):
-    """
-    Base class to provide source management: discovery, registration
-    Actual launch of tasks must be defined in subclasses
-    """
-
-    # define the class manager will look for. Set in a subclass
-    SOURCE_CLASS = None
-
-    def __init__(self, job_manager, datasource_path="dataload.sources", *args, **kwargs):
-        super(BaseSourceManager, self).__init__(job_manager, *args, **kwargs)
-        self.conn = get_src_conn()
-        self.default_src_path = datasource_path
-        self.poll_schedule = None
-
-    def filter_class(self, klass):
-        """
-        Gives opportunity for subclass to check given class and decide to
-        keep it or not in the discovery process. Returning None means "skip it".
-        """
-        # keep it by default
-        return klass
-
-    def register_classes(self, klasses):
-        """
-        Register each class in self.register dict. Key will be used
-        to retrieve the source class, create an instance and run method from it.
-        It must be implemented in subclass as each manager may need to access
-        its sources differently,based on different keys.
-        """
-        raise NotImplementedError("implement me in sub-class")
-
-    def find_classes(self, src_module, fail_on_notfound=True):
-        """
-        Given a python module, return a list of classes in this module, matching
-        SOURCE_CLASS (must inherit from)
-        """
-        # try to find a uploader class in the module
-        klasses = []
-        for attr in dir(src_module):
-            something = getattr(src_module, attr)
-            # not interested in classes coming from biothings.hub.*, these would typically come
-            # from "from biothings.hub.... import aclass" statements and would be incorrectly registered
-            # we only look for classes defined straight from the actual module
-            if (
-                type(something) == type
-                and issubclass(something, self.__class__.SOURCE_CLASS)
-                and not something.__module__.startswith("biothings.hub")
-            ):
-                klass = something
-                if not self.filter_class(klass):
-                    continue
-                logger.debug("Found a class based on %s: '%s'", self.__class__.SOURCE_CLASS.__name__, klass)
-                klasses.append(klass)
-        if not klasses:
-            if fail_on_notfound:
-                raise UnknownResource(
-                    "Can't find a class based on %s in module '%s'" % (self.__class__.SOURCE_CLASS.__name__, src_module)
-                )
-        return klasses
-
-    def register_source(self, src, fail_on_notfound=True):
-        """Register a new data source. src can be a module where some classes
-        are defined. It can also be a module path as a string, or just a source name
-        in which case it will try to find information from default path.
-        """
-        if isinstance(src, str):
-            try:
-                src_m = importlib.import_module(src)
-                src_m = importlib.reload(src_m)
-            except ImportError as ex:
-                logger.warning(ex)
-                try:
-                    src_m = importlib.import_module("%s.%s" % (self.default_src_path, src))
-                except ImportError as e:
-                    msg = "Can't find module '%s', even in '%s'" % (src, self.default_src_path)
-                    logger.error(msg)
-                    logger.warning(e)
-                    raise UnknownResource(msg)
-
-        elif isinstance(src, dict):
-            # source is comprised of several other sub sources
-            assert len(src) == 1, "Should have only one element in source dict '%s'" % src
-            _, sub_srcs = list(src.items())[0]
-            for src in sub_srcs:
-                self.register_source(src, fail_on_notfound)
-            return
-        else:
-            src_m = src
-        # first try to find classes defined in __init__.py (in package) explicitely
-        klasses = self.find_classes(src_m, fail_on_notfound)
-        # then if none found, try to search within the package's modules
-        if not klasses:
-            try:
-                src_m_path = src_m.__path__[0]
-                for d in os.listdir(src_m_path):
-                    if d.endswith("__pycache__"):
-                        continue
-                    modpath = os.path.join(src_m.__name__, d).replace(".py", "").replace(os.path.sep, ".")
-                    try:
-                        m = importlib.import_module(modpath)
-                        klasses.extend(self.find_classes(m, fail_on_notfound))
-                    except Exception as e:
-                        # (SyntaxError, ImportError) is not sufficient to catch
-                        # all possible failures, for example a ValueError
-                        # in module definition..
-                        logger.debug("Couldn't import %s: %s", modpath, e)
-                        continue
-            except TypeError as e:
-                logger.warning("Can't register source '%s', something's wrong with path: %s", src_m, e)
-        logger.debug("Found classes to register: %s", repr(klasses))
-
-        self.register_classes(klasses)
-
-    def register_sources(self, sources):
-        assert not isinstance(sources, str), "sources argument is a string, should pass a list"
-        self.register.clear()
-        for src in sources:
-            try:
-                # batch registration, we'll silently ignore not-found sources
-                self.register_source(src, fail_on_notfound=False)
-            except (UnknownResource, ResourceError) as e:
-                logger.info("Can't register source '%s', skip it; %s", src, e)
-                import traceback
-
-                logger.error(traceback.format_exc())
 
 
 class JobManager:
